# static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Documentation Field Reservation System",
    "description": "by Rofif Fairuz Hawary\n\nLast Updated: 23/04/2024 00:03:XX",
    "contact": {
      "email": "rofifhawary@gmail.com"
    },
    "version": "1.7.7"
  },
...
  "servers": [
    {
      "url": "http://rofif.my.id/"
    },
    {
      "url": "http://localhost:5000/"
    }
  ],
...
  "paths": {
...
    "/player/sportVenue": {
      "post": {
        "tags": [
          "playerVenueInfo"
        ],
        "summary": "Return all of Venue list based on filter and sorting",
        "description": "Return all of Venue list based on filter and sorting",
        "operationId": "playerGetSportVenue",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Fill field with null values if not used, there are 2 sort kinds ('distance' or 'price')",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerGetSportVenueListBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerGetSportVenueListResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid input"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/player/sportVenue/{venue_id}": {
      "post": {
        "tags": [
          "playerVenueInfo"
        ],
        "summary": "Get detailed venue info by venue_id",
        "description": "",
        "operationId": "playerGetSportVenueDetail",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "header",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "coordinate = pin coordinate player",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerGetSportVenueDetailBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerGetVenueDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid input"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/player/sportVenue/fields/{Sport_Venue_id}": {
      "get": {
        "tags": [
          "playerVenueInfo"
        ],
        "summary": "Get List field from managed sport venue by admin",
        "description": "Get List field from managed sport venue by admin",
        "operationId": "playerGetFieldFromSportVenue",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Sport_Venue_id",
            "in": "path",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerGetFieldFromVenueResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid input"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/player/sportVenue/fields/schedule/blacklist/{field_id}/{month}/{year}": {
      "get": {
        "tags": [
          "playerVenueInfo"
        ],
        "summary": "Get Blacklisted schedule in a month (with parameter year and month)",
        "description": "Get Blacklisted schedule in a month (with parameter year and month)",
        "operationId": "playerGetBlacklistSchedule",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "in": "path",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerGetBlacklistScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid input"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/player/sportVenue/fields/schedule/reservation/{field_id}/{month}/{year}": {
      "get": {
        "tags": [
          "playerVenueInfo"
        ],
        "summary": "Get Reservation schedule in a month (with parameter year and month)",
        "description": "Get Reservation schedule in a month (with parameter year and month)",
        "operationId": "playerGetReservationSchedule",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "in": "path",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetReservationScheduleFieldsResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid input"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PlayerGetFieldFromVenueResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "get_status": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "Sport_Field_id": {
                  "type": "string"
                },
                "number": {
                  "type": "integer"
                }
              }
            }
          }
        },
        "xml": {
          "name": "PlayerGetFieldFromVenueResponse"
        }
      },
...
      "PlayerGetVenueDetailResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "get_status": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "Sport_Kind_id": {
                "type": "string"
              },
              "Sport_Kind_Name": {
                "type": "string"
              },
              "created_at": {
                "type": "string"
              },
              "last_edited": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "rules": {
                "type": "string"
              },
              "geo_coordinate": {
                "type": "string"
              },
              "distance": {
                "type": "float"
              },
              "is_bike_parking": {
                "type": "boolean"
              },
              "is_car_parking": {
                "type": "boolean"
              },
              "is_public": {
                "type": "boolean"
              },
              "name": {
                "type": "string"
              },
              "time_open": {
                "type": "string"
              },
              "time_closed": {
                "type": "string"
              },
              "price_per_hour": {
                "type": "integer"
              }
            }
          }
        },
        "xml": {
          "name": "PlayerGetVenueDetailResponse"
        }
      },
...
      "PlayerGetBlacklistScheduleResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "get_status": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "blacklist_id": {
                  "type": "string"
                },
                "date": {
                  "type": "string"
                },
                "fromTime": {
                  "type": "string"
                },
                "toTime": {
                  "type": "string"
                },
                "reason": {
                  "type": "string"
                }
              }
            }
          }
        },
        "xml": {
          "name": "PlayerGetBlacklistScheduleResponse"
        }
      },
      "GetReservationScheduleFieldsResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "get_status": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "Field_id": {
                  "type": "string"
                },
                "Player_username": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "mabar_type": {
                  "type": "string"
                },
                "date": {
                  "type": "string"
                },
                "time_start": {
                  "type": "string"
                },
                "time_end": {
                  "type": "string"
                },
                "booking_status": {
                  "type": "string"
                },
                "payment_credential_url": {
                  "type": "string"
                },
                "is_public": {
                  "type": "string"
                },
                "is_open_member": {
                  "type": "string"
                }
              }
            }
          }
        },
        "xml": {
          "name": "GetReservationScheduleFieldsResponse"
        }
      },
...
      "PlayerGetSportVenueListBody": {
        "type": "object",
        "properties": {
          "is_car_parking": {
            "type": "boolean",
            "example": false
          },
          "is_bike_parking": {
            "type": "boolean",
            "example": true
          },
          "name": {
            "type": "string",
            "example": null
          },
          "Sport_Kind_id": {
            "type": "string",
            "example": null
          },
          "sort_by": {
            "type": "string",
            "example": "distance"
          },
          "coordinate": {
            "type": "string",
            "example": "-6.894797, 107.610590"
          }
        },
        "xml": {
          "name": "PlayerGetSportVenueListBody"
        }
      },
      "PlayerGetSportVenueDetailBody": {
        "type": "object",
        "properties": {
          "coordinate": {
            "type": "string",
            "example": "-6.894797, 107.610590"
          }
        },
        "xml": {
          "name": "PlayerGetSportVenueDetailBody"
        }
      },
      "PlayerGetSportVenueListResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "get_status": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "Sport_Kind_Name": {
                  "type": "string"
                },
                "geo_coordinate": {
                  "type": "string"
                },
                "distance": {
                  "type": "float"
                },
                "is_bike_parking": {
                  "type": "boolean"
                },
                "is_car_parking": {
                  "type": "boolean"
                },
                "is_public": {
                  "type": "boolean"
                },
                "name": {
                  "type": "string"
                },
                "price_per_hour": {
                  "type": "integer"
                }
              }
            }
          }
        },
        "xml": {
          "name": "PlayerGetSportVenueListResponse"
        }
      },
...
    }
  }
}